cleanup and added a beginning mockito test to start building from
