+++ build.sbt
@@ -3,7 +3,8 @@ lazy val akkaVersion    = "2.6.14"
 lazy val logBackVersion = "1.2.3"
 lazy val scalaTestVersion = "3.1.4"
 lazy val mongoScalaDriver = "4.2.3"
-lazy val fakeMongo = "2.1.0"
+lazy val mockitoVersion = "3.9.0"
+lazy val mockitoScalaTest = "1.16.37"
 
 lazy val root = (project in file(".")).
   settings(
@@ -26,7 +27,8 @@ lazy val root = (project in file(".")).
       "com.typesafe.akka" %% "akka-http-testkit"        % akkaHttpVersion   % Test,
       "com.typesafe.akka" %% "akka-actor-testkit-typed" % akkaVersion       % Test,
       "org.scalatest"     %% "scalatest"                % scalaTestVersion  % Test,
-      "org.mongodb.scala" %% "mongo-scala-driver"       % mongoScalaDriver  % Test,
-      "com.github.fakemongo" % "fongo"                  % fakeMongo         % Test
+      "org.mockito"       % "mockito-core"              % mockitoVersion    % Test,
+      "org.mockito"       %% "mockito-scala-scalatest"  % mockitoScalaTest  % Test,
+      "org.mongodb.scala" %% "mongo-scala-driver"       % mongoScalaDriver  % Test
     )
   )
