--- project/plugins.sbt
@@ -1,2 +1,0 @@
-addSbtPlugin("com.waioeka.sbt" % "cucumber-plugin" % "0.1.7")
-
